Move source files in aws-appsync-api from aws.api.appsync to aws.api package
