+++ aws-api-appsync/src/main/java/com/amplifyframework/api/aws/AppSyncGraphQLRequest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.api.aws.appsync;
+package com.amplifyframework.api.aws;
 
 import android.text.TextUtils;
 import androidx.annotation.NonNull;
