+++ aws-api-appsync/src/main/java/com/amplifyframework/api/aws/GsonVariablesSerializer.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.api.aws.appsync;
+package com.amplifyframework.api.aws;
 
 import com.amplifyframework.api.graphql.GraphQLRequest;
 import com.amplifyframework.core.model.temporal.Temporal;
