+++ aws-api-appsync/src/test/java/com/amplifyframework/api/aws/DefaultGraphQLRequestOptions.java
@@ -13,7 +13,9 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.api.aws.appsync;
+package com.amplifyframework.api.aws;
+
+import com.amplifyframework.api.aws.GraphQLRequestOptions;
 
 import java.util.Arrays;
 import java.util.Collections;
