+++ aws-api/src/main/java/com/amplifyframework/api/aws/AppSyncPaginatedResultDeserializer.java
@@ -15,7 +15,6 @@
 
 package com.amplifyframework.api.aws;
 
-import com.amplifyframework.api.aws.appsync.AppSyncGraphQLRequest;
 import com.amplifyframework.api.graphql.PaginatedResult;
 import com.amplifyframework.core.model.Model;
